+++ website/scully.typedb-web.config.ts
@@ -1,6 +1,30 @@
-import { ScullyConfig } from "@scullyio/scully";
+import { getSitemapPlugin, SitemapConfig } from "@gammastream/scully-plugin-sitemap";
+import { ScullyConfig, setPluginConfig } from "@scullyio/scully";
 import "@scullyio/scully-plugin-puppeteer";
-import { applicationArticleRoutes, blogCategoryRoutes, blogPostRoutes, eventRoutes, fundamentalArticleRoutes, lectureRoutes, whitePaperRoutes } from "./scully/plugins/plugins";
+
+import {
+    applicationArticleRoutes,
+    blogCategoryRoutes,
+    blogPostRoutes,
+    eventRoutes,
+    fundamentalArticleRoutes,
+    lectureRoutes,
+    whitePaperRoutes,
+} from "./scully/plugins/plugins";
+
+const SitemapPlugin = getSitemapPlugin();
+setPluginConfig<SitemapConfig>(SitemapPlugin, {
+    urlPrefix: process.env["URL"] || "https://typedb.com",
+    changeFreq: "daily",
+    sitemapFilename: "sitemap-main.xml",
+    ignoredRoutes: ["/__404"],
+    routes: {
+        "/": { priority: "0.9" },
+        "/philosophy": { priority: "0.8" },
+        "/features": { priority: "0.7" },
+        "/cloud": { priority: "0.6" },
+    },
+});
 
 export const config: ScullyConfig = {
     projectRoot: "./src",
@@ -16,6 +40,7 @@ export const config: ScullyConfig = {
         "/white-papers/:slug": { type: whitePaperRoutes },
         "/events/:slug": { type: eventRoutes },
     },
+    extraRoutes: ["/__404"],
     puppeteerLaunchOptions: {
         // executablePath: "/opt/homebrew/bin/chromium",
         args: ["--no-sandbox", "--disabled-setupid-sandbox"],
