+++ website/scully/plugins/plugins.ts
@@ -1,5 +1,4 @@
-import { getSitemapPlugin, SitemapConfig } from "@gammastream/scully-plugin-sitemap";
-import { HandledRoute, registerPlugin, setPluginConfig } from "@scullyio/scully";
+import { HandledRoute, registerPlugin } from "@scullyio/scully";
 import axios from "axios";
 
 const defaultValidator = async () => [];
@@ -71,17 +70,3 @@ registerPlugin("router", blogPostRoutes, blogPostRoutesPlugin, defaultValidator)
 registerPlugin("router", lectureRoutes, lectureRoutesPlugin, defaultValidator);
 registerPlugin("router", whitePaperRoutes, whitePaperRoutesPlugin, defaultValidator);
 registerPlugin("router", eventRoutes, eventRoutesPlugin, defaultValidator);
-
-const SitemapPlugin = getSitemapPlugin();
-setPluginConfig<SitemapConfig>(SitemapPlugin, {
-    urlPrefix: process.env.URL || "https://typedb.com",
-    changeFreq: "daily",
-    sitemapFilename: "sitemap-main.xml",
-    ignoredRoutes: ["/404"],
-    routes: {
-        "/": { priority: "0.9" },
-        "/philosophy": { priority: "0.8" },
-        "/features": { priority: "0.7" },
-        "/cloud": { priority: "0.6" },
-    },
-});
